removes the location service to its utility class
